% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ghypergeo}
\alias{ghypergeo}
\title{Generalized hypergeometric function}
\usage{
ghypergeo(a, b, z, k)
}
\arguments{
\item{a}{matrix of upper hypergeometric parameters.}

\item{b}{matrix of lower hypergeometric parameters.}

\item{z}{vector of hypergeometric function arguments.}

\item{k}{maximum number of hypergeometric series terms. Non-positive values
use the package default.}
}
\value{
A one-column numeric matrix of generalized hypergeometric function
values. Rows of \code{a}, rows of \code{b}, and values of \code{z} are
recycled to the output length.
}
\description{
Evaluate generalized hypergeometric series used by the split-t gradient
calculations.
}
\seealso{
\code{\link{gsplitt}()}
}
