\name{rcpp_splitt}
\alias{rcpp_splitt}
\title{Splitt}
\usage{
dsplitt(x, mu = 0, df, phi, lmd,  log = FALSE)
}
\description{
Density, distribution function, quantile function and random generation for the t distribution with mean equal to mean and standard deviation equal to sd.

}
\examples{
dsplitt(1,0,1,1,1,F)
}
