% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsplitn.R
\name{qsplitn}
\alias{qsplitn}
\title{the quantile function of split-normal}
\usage{
qsplitn(p, mu, sigma, lmd)
}
\arguments{
\item{p}{vector of probabilities.}

\item{mu}{vector of means.}

\item{sigma}{vector of standard deviations.}

\item{lmd}{vector of skewness.}
}
\value{
vector of quantile
}
